[HTTP]: Add errorLoger to all posible errors

+++ http-server/src/controllers/contactsController.ts
@@ -1,4 +1,4 @@
-import { contactsAPI } from 'database';
+import { contactsAPI, errorLogManager } from 'database';
 
 import type { Request, Response } from 'express';
 
@@ -11,7 +11,11 @@ class ContactsController {
 
       res.status(200).send({ message: contacts });
     } catch (error) {
-      res.status(500).json({ message: (error as Error).message });
+      errorLogManager.addToLogs(
+        'Error in ChatController in get',
+        `${JSON.stringify(req.body)}`,
+      );
+      return res.status(500).json({ message: (error as Error).message });
     }
   }
   async post(req: Request, res: Response) {
@@ -21,15 +25,38 @@ class ContactsController {
 
       const contact = await contactsAPI.getById(Number(id), Number(contactId));
 
+      if (!contact) {
+        errorLogManager.addToLogs(
+          'Error in ContactsController in post: contact === null',
+          `${JSON.stringify(req.body)}`,
+        );
+        return res.status(403).json({ message: 'Invalid access' });
+      }
+
       if (contact!.length > 0) {
+        errorLogManager.addToLogs(
+          'Error in ContactsController in post: double contact add',
+          `${JSON.stringify(req.body)}`,
+        );
         return res.status(403).json({ message: 'Invalid access' });
       }
 
       const contacts = await contactsAPI.post(Number(id), contactId);
 
+      if (!contacts) {
+        errorLogManager.addToLogs(
+          'Error in ContactsController in post. contacts === null',
+          `${JSON.stringify(req.body)}`,
+        );
+        return res.status(500).json({ message: 'Error post' });
+      }
       res.status(200).send({ message: contacts });
     } catch (error) {
-      res.status(500).json({ message: (error as Error).message });
+      errorLogManager.addToLogs(
+        'Error in ContactsController in post',
+        `${JSON.stringify(req.body)}`,
+      );
+      return res.status(500).json({ message: (error as Error).message });
     }
   }
   async del(req: Request, res: Response) {
@@ -38,14 +65,34 @@ class ContactsController {
 
       const contact = await contactsAPI.getById(Number(id), Number(contactId));
 
+      if (!contact) {
+        errorLogManager.addToLogs(
+          'Error in ContactsController in del: contact === null',
+          `${JSON.stringify(req.body)}`,
+        );
+        return res.status(403).json({ message: 'Invalid access' });
+      }
+
       if (contact!.length === 0) {
         return res.status(403).json({ message: 'Invalid access' });
       }
 
       const contacts = await contactsAPI.delete(Number(id), Number(contactId));
 
+      if (!contacts) {
+        errorLogManager.addToLogs(
+          'Error in ContactsController in del. contacts === null',
+          `${JSON.stringify(req.body)}`,
+        );
+        return res.status(500).json({ message: 'Error post' });
+      }
+
       res.status(200).send({ message: contacts });
     } catch (error) {
+      errorLogManager.addToLogs(
+        'Error in ContactsController in del',
+        `${(error as Error).message}`,
+      );
       res.status(500).json({ message: (error as Error).message });
     }
   }
